large channel update

+++ libOTe_Tests/OTOracleReceiver.cpp
@@ -30,7 +30,7 @@ namespace tests_libOTe
 
         std::array<block, 2> ss;
 
-        for (u64 doneIdx = 0; doneIdx < messages.size(); ++doneIdx)
+        for (u64 doneIdx = 0; doneIdx < u64(messages.size()); ++doneIdx)
         {
             ss[0] = mPrng.get<block>();
             ss[1] = mPrng.get<block>();
